Rename treewalkers._base to .base to reflect public status
