+++ html5lib/tests/test_serializer.py
@@ -10,7 +10,7 @@ from .support import get_data_files
 from html5lib import constants
 from html5lib.filters.lint import Filter as Lint
 from html5lib.serializer import HTMLSerializer, serialize
-from html5lib.treewalkers._base import TreeWalker
+from html5lib.treewalkers.base import TreeWalker
 
 # pylint:disable=wrong-import-position
 optionals_loaded = []
