+++ html5lib/treewalkers/base.py
